move json encoder tests
